refactor: Moved bunch of stuff out of `AliasUtil`
`AliasUtil` had grown to be somewhat of a kitchen sink. So this
refactor moves things that belong to aliases into `Alias`, things
having to do with catalogs into `Catalog` etc.
