+++ src/main/java/dev/jbang/source/Source.java
@@ -7,7 +7,6 @@ import java.util.Optional;
 import java.util.Properties;
 
 import dev.jbang.catalog.Alias;
-import dev.jbang.catalog.AliasUtil;
 import dev.jbang.cli.BaseCommand;
 import dev.jbang.cli.ExitException;
 import dev.jbang.dependencies.ModularClassPath;
@@ -118,7 +117,7 @@ public interface Source {
 		Alias alias = null;
 		if (resourceRef == null) {
 			// Not found as such, so let's check the aliases
-			alias = AliasUtil.getAlias(null, resource, ctx.getArguments(), ctx.getProperties());
+			alias = Alias.get(null, resource, ctx.getArguments(), ctx.getProperties());
 			if (alias != null) {
 				resourceRef = ResourceRef.forResource(alias.resolve(null));
 				ctx.setArguments(alias.arguments);
