+++ src/test/java/dev/jbang/util/TestUtil.java
@@ -15,11 +15,11 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import dev.jbang.BaseTest;
-import dev.jbang.catalog.AliasUtil;
+import dev.jbang.catalog.Catalog;
 
 public class TestUtil extends BaseTest {
 	public static void clearSettingsCaches() {
-		AliasUtil.clearCache();
+		Catalog.clearCache();
 	}
 
 	@Test
